Updated direct route names
