+++ src/app/routes/api/account_info.py
@@ -9,7 +9,7 @@ from flask_login import login_required
 from Crypto.Cipher import PKCS1_OAEP
 from Crypto.PublicKey import RSA
 from Crypto.Hash import SHA256
-from ...database.database2 import (
+from ...database.database import (
     Database,
     DatabaseException,
     DuplicateUserException,
