+++ src/app/routes/api/global_ingredients.py
@@ -2,7 +2,7 @@
 This file contains endpoints related to global ingredient data.
 """
 from flask import Blueprint, request, Flask
-from ...database.database2 import (
+from ...database.database import (
     Database,
     DatabaseException,
     NoIngredientException,
