+++ src/app/routes/api/users.py
@@ -2,7 +2,7 @@
 This file contains endpoints related to users.
 """
 from flask import Blueprint, request, Flask
-from ...database.database2 import (
+from ...database.database import (
     Database,
     DatabaseException,
     InvalidArgumentException,
