+++ src/app/routes/html/account.py
@@ -1,17 +1,20 @@
 """
-This file contains user-facing endpoints relating to profile pages.
+This file contains user-facing endpoints relating to account pages.
 """
 from flask import Flask, Blueprint, render_template
 from flask_login import login_required
-from ...database.database2 import (
+from ...database.database import (
     Database,
+    DatabaseException,
 )
 from ... import util
-from ..routing_util import get_current_user
-from ...api import spoonacular
+from ..routing_util import (
+    get_current_user,
+    NoCurrentUserException,
+)
 
 blueprint = Blueprint(
-    "bp_direct_profile",
+    "bp_direct_account",
     __name__,
     template_folder=util.get_templates_folder(),
     static_folder=util.get_static_folder(),
@@ -50,27 +53,21 @@ def init(app: Flask, database: Database):
     app.register_blueprint(get_blueprint())
 
 
-@blueprint.route("/profile")
+@blueprint.route("/account")
 @login_required
-def profile():
+def account():
     """
-    Returns the profile page.
+    Returns the account page.
     """
-    current_user = get_current_user()
-    (recipes, _) = DATABASE.get_recipes(current_user.id)
-    (ingredients, _) = DATABASE.get_ingredients(current_user.id)
-
-    if len(recipes) == 0:
-        try:
-            recipes = spoonacular.get_recommended_recipes()
-        except spoonacular.SpoonacularApiException:
-            pass
-
-    return render_template(
-        "profile.html",
-        recipes=[recipe.to_json() for recipe in recipes],
-        has_recipes=len(recipes) > 0,
-        ingredients=[ingredient.to_json() for ingredient in ingredients],
-        has_ingredient=len(ingredients) > 0,
-        userdata=current_user.to_json(shallow=True),
-    )
+    try:
+        current_user = get_current_user()
+        userdata = current_user.to_json()
+        userdata["friends"] = [
+            x.to_json(shallow=True)
+            for x in DATABASE.get_relationships_for_user(current_user.id)
+        ]
+        return render_template("account.html", userdata=userdata)
+    except DatabaseException:
+        return None  # TODO: Handle exceptions
+    except NoCurrentUserException:
+        return None  # TODO: Handle exceptions
