+++ src/app/routes/html/html_routes.py
@@ -1,17 +1,17 @@
 """
 This file is responsible for collecting all of the
-direct routes into one file so the application need only
+HTML routes into one file so the application need only
 call init() from this module.
 """
 
 from flask import Flask
-from ...database.database2 import Database
+from ...database.database import Database
 from . import account, login, signup, index, profile
 
 
 def init(app: Flask, database: Database):
     """
-    Initializes all direct route modules.
+    Initializes all HTML route modules.
     """
     account.init(app, database)
     login.init(app, database)
