+++ src/app/routes/html/index.py
@@ -3,7 +3,7 @@ This file contains user-facing endpoints relating to index pages.
 """
 import random
 from flask import Flask, Blueprint, render_template
-from ...database.database2 import (
+from ...database.database import (
     Database,
 )
 from ... import util
